server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
Pillow==10.1.0
openai==1.3.7
pyyaml==6.0.1
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
matplotlib==3.7.2
open3d==0.17.0 